backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic>=2.10.0
anthropic>=0.79.0
openai>=1.60.0
python-dotenv>=1.0.0
